generic in c++
in this we write a small code using generics in c++
generic programming basically enables the programmer to write a general algorithm which will work  in all data types
Write a single generic function named printArray; this function must take an array of generic elements as a parameter (the exception to this is C++, which takes a vector). The locked Solution class in your editor tests your function.